# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: API specification para Sistema de Gestion
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /ventas:
    get:
      operationId: getVentas
      tags: [ventas]
      summary: Obtener ventas (por fecha o todas)
      parameters:
        - name: fecha
          in: query
          required: false
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Lista de ventas del dia
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/VentaDiaria"
    post:
      operationId: crearVenta
      tags: [ventas]
      summary: Registrar una venta diaria
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VentaDiariaInput"
      responses:
        "201":
          description: Venta registrada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VentaDiaria"
  /ventas/{id}:
    put:
      operationId: actualizarVenta
      tags: [ventas]
      summary: Actualizar una venta
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VentaDiariaInput"
      responses:
        "200":
          description: Venta actualizada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VentaDiaria"
    delete:
      operationId: eliminarVenta
      tags: [ventas]
      summary: Eliminar una venta
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Eliminado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MensajeRespuesta"
...
  /manoobra:
    get:
      operationId: getManoObra
      tags: [manoobra]
      summary: Obtener registros de mano de obra
      parameters:
        - name: fecha
          in: query
          required: false
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Lista de mano de obra
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ManoObra"
    post:
      operationId: crearManoObra
      tags: [manoobra]
      summary: Registrar mano de obra
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ManoObraInput"
      responses:
        "201":
          description: Mano de obra registrada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ManoObra"
  /manoobra/{id}:
    put:
      operationId: actualizarManoObra
      tags: [manoobra]
      summary: Actualizar mano de obra
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ManoObraInput"
      responses:
        "200":
          description: Mano de obra actualizada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ManoObra"
    delete:
      operationId: eliminarManoObra
      tags: [manoobra]
      summary: Eliminar mano de obra
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Eliminado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MensajeRespuesta"
  /trabajadores:
    get:
      operationId: getTrabajadores
      tags: [trabajadores]
      summary: Obtener lista de trabajadores
      responses:
        "200":
          description: Lista de trabajadores
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Trabajador"
    post:
      operationId: crearTrabajador
      tags: [trabajadores]
      summary: Crear trabajador
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TrabajadorInput"
      responses:
        "201":
          description: Trabajador creado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trabajador"
  /trabajadores/{id}:
    put:
      operationId: actualizarTrabajador
      tags: [trabajadores]
      summary: Actualizar trabajador
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TrabajadorInput"
      responses:
        "200":
          description: Trabajador actualizado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trabajador"
...
components:
  schemas:
...
    MensajeRespuesta:
      type: object
      properties:
        mensaje:
          type: string
      required:
        - mensaje
...
    VentaDiaria:
      type: object
      properties:
        id:
          type: integer
        fecha:
          type: string
          format: date
        referencia:
          type: string
          description: "No. Remision, placa, nombre cliente o descripcion"
        tipoLinea:
          type: string
          enum: [venta, manoobra, credito]
          description: "venta=normal, manoobra=amarillo, credito=azul"
        productoId:
          type: integer
          nullable: true
        productoNombre:
          type: string
          nullable: true
        productoCodigo:
          type: string
          nullable: true
        productoMarca:
          type: string
          nullable: true
        cantidad:
          type: number
        precioCompraUnidad:
          type: number
        precioVentaUnidad:
          type: number
        precioVentaTotal:
          type: number
          description: precioVentaUnidad * cantidad (ya incluye IVA si aplica)
        beneficio:
          type: number
          description: Beneficio de la venta
        descripcion:
          type: string
          nullable: true
          description: Descripcion adicional (para manoobra o credito)
        creadoEn:
          type: string
          format: date-time
      required:
        - id
        - fecha
        - referencia
        - tipoLinea
        - cantidad
        - precioCompraUnidad
        - precioVentaUnidad
        - precioVentaTotal
        - beneficio
    VentaDiariaInput:
      type: object
      properties:
        fecha:
          type: string
          format: date
        referencia:
          type: string
        tipoLinea:
          type: string
          enum: [venta, manoobra, credito]
        productoId:
          type: integer
          nullable: true
        productoNombre:
          type: string
          nullable: true
        productoCodigo:
          type: string
          nullable: true
        productoMarca:
          type: string
          nullable: true
        cantidad:
          type: number
        precioCompraUnidad:
          type: number
        precioVentaUnidad:
          type: number
        precioVentaTotal:
          type: number
        beneficio:
          type: number
        descripcion:
          type: string
          nullable: true
      required:
        - fecha
        - referencia
        - tipoLinea
        - cantidad
        - precioCompraUnidad
        - precioVentaUnidad
        - precioVentaTotal
        - beneficio
...
    ManoObra:
      type: object
      properties:
        id:
          type: integer
        fecha:
          type: string
          format: date
        descripcion:
          type: string
        valorTotal:
          type: number
          description: Valor total cobrado al cliente
        distribuciones:
          type: array
          items:
            $ref: "#/components/schemas/DistribucionTrabajador"
        creadoEn:
          type: string
          format: date-time
      required:
        - id
        - fecha
        - descripcion
        - valorTotal
        - distribuciones
    ManoObraInput:
      type: object
      properties:
        fecha:
          type: string
          format: date
        descripcion:
          type: string
        valorTotal:
          type: number
        distribuciones:
          type: array
          items:
            $ref: "#/components/schemas/DistribucionTrabajadorInput"
      required:
        - fecha
        - descripcion
        - valorTotal
        - distribuciones
    DistribucionTrabajador:
      type: object
      properties:
        id:
          type: integer
        trabajadorId:
          type: integer
        trabajadorNombre:
          type: string
        valor:
          type: number
        descuentoSeguro:
          type: number
        descuentoOtros:
          type: number
      required:
        - id
        - trabajadorId
        - trabajadorNombre
        - valor
        - descuentoSeguro
        - descuentoOtros
    DistribucionTrabajadorInput:
      type: object
      properties:
        trabajadorId:
          type: integer
        valor:
          type: number
        descuentoSeguro:
          type: number
        descuentoOtros:
          type: number
      required:
        - trabajadorId
        - valor
        - descuentoSeguro
        - descuentoOtros
    Trabajador:
      type: object
      properties:
        id:
          type: integer
        nombre:
          type: string
        descuentoSeguro:
          type: number
          description: Descuento mensual de seguro (20000 para trabajadores 1 y 2)
        descuentoOtros:
          type: number
          description: Descuento por compras de la empresa (trabajador 3)
        totalGanado:
          type: number
          description: Total acumulado ganado
        totalDescuentos:
          type: number
          description: Total descuentos aplicados
        activo:
          type: boolean
        creadoEn:
          type: string
          format: date-time
      required:
        - id
        - nombre
        - descuentoSeguro
        - descuentoOtros
        - totalGanado
        - totalDescuentos
        - activo
    TrabajadorInput:
      type: object
      properties:
        nombre:
          type: string
        descuentoSeguro:
          type: number
        descuentoOtros:
          type: number
        activo:
          type: boolean
      required:
        - nombre
        - descuentoSeguro
        - descuentoOtros
...