# lib/api-zod/src/generated/types/manoObraInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification para Sistema de Gestion
 * OpenAPI spec version: 0.1.0
 */
import type { DistribucionTrabajadorInput } from "./distribucionTrabajadorInput";

export interface ManoObraInput {
  fecha: Date;
  descripcion: string;
  valorTotal: number;
  distribuciones: DistribucionTrabajadorInput[];
}