# lib/api-zod/src/generated/types/trabajadorInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification para Sistema de Gestion
 * OpenAPI spec version: 0.1.0
 */

export interface TrabajadorInput {
  nombre: string;
  descuentoSeguro: number;
  descuentoOtros: number;
  activo?: boolean;
}