# lib/api-zod/src/generated/types/ventaDiaria.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification para Sistema de Gestion
 * OpenAPI spec version: 0.1.0
 */
import type { VentaDiariaTipoLinea } from "./ventaDiariaTipoLinea";

export interface VentaDiaria {
  id: number;
  fecha: Date;
  /** No. Remision, placa, nombre cliente o descripcion */
  referencia: string;
  /** venta=normal, manoobra=amarillo, credito=azul */
  tipoLinea: VentaDiariaTipoLinea;
  productoId?: number | null;
  productoNombre?: string | null;
  productoCodigo?: string | null;
  productoMarca?: string | null;
  cantidad: number;
  precioCompraUnidad: number;
  precioVentaUnidad: number;
  /** precioVentaUnidad * cantidad (ya incluye IVA si aplica) */
  precioVentaTotal: number;
  /** Beneficio de la venta */
  beneficio: number;
  /** Descripcion adicional (para manoobra o credito) */
  descripcion?: string | null;
  creadoEn?: Date;
}